Reorganizing component tests
